Added try/catch around decryption in FileSavable
When getting file contents, if encryption/decryption is enabled and fails decryption, the file will be returned from getFileContents() as it exists on disk. This was added so that the $encryptFile flag can be toggled back and forth in the model and we should still be able to access any files that already exist on the disk, regardless of whether or not they are already encrypted.